Reorganize directory for organization list
